+++ CHANGES.rst
@@ -37,6 +37,9 @@ Bug Fixes
 
 - ``astropy.io.ascii``
 
+  - Fix a bug in Python 3 when guessing file format using a file object as
+    input.  Also improve performance in same situation for Python 2. [#3132]
+
 - ``astropy.io.fits``
 
   - The ``in`` operator now works correctly for checking if an extension